// old/agir-work-planning-web/src/app/map/config/layers/logic-layers/map-logic-layer-enum.ts
export enum MapLogicLayer {
  addresses = 'addresses',
  aqueductAccessory = 'aqueductAccessory',
  assetsPins = 'assetsPins',
  basemapLabels = 'basemapLabels',
  boroughs = 'boroughs',
  boroughsPolygons = 'boroughsPolygons',
  busShelter = 'busShelter',
  countByBorough = 'countByBorough',
  countByCity = 'countByCity',
  fireHydrants = 'fireHydrant',
  futureProjectAreas = 'futureProjectAreas',
  futureRtuProjectAreas = 'futureRtuProjectAreas',
  futureProjectPins = 'futureProjectPins',
  greenSpace = 'greenSpace',
  preliminaryOrderedProjectPins = 'preliminaryOrderedProjectPins',
  canceledProjectPins = 'canceledProjectPins',
  plannedProjectPins = 'plannedProjectPins',
  postponedProjectPins = 'postponedProjectPins',
  finalOrderedProjectPins = 'finalOrderedProjectPins',
  replannedProjectPins = 'replannedProjectPins',
  programmedProjectPins = 'programmedProjectPins',
  futureRtuProjectPins = 'futureRtuProjectPins',
  greyBasemap = 'greyBasemap',
  highways = 'highway',
  interventionCreation = 'interventionCreation',
  interventions = 'interventions',
  leadGround = 'leadGround',
  mobilityAxis = 'mobilityAxis',
  multipleYearsProjectAreas = 'multipleYearsProjectAreas',
  multipleYearsPlannedProjectAreas = 'multipleYearsPlannedProjectAreas',
  multipleYearsReplannedProjectAreas = 'multipleYearsReplannedProjectAreas',
  multipleYearsPostponedProjectAreas = 'multipleYearsPostponedProjectAreas',
  pastProjectAreas = 'pastProjectAreas',
  pastRtuProjectAreas = 'pastRtuProjectAreas',
  pastProjectPins = 'pastProjectPins',
  pastRtuProjectPins = 'pastRtuProjectPins',
  presentProjectAreas = 'presentProjectAreas',
  plannedProjectAreas = 'plannedProjectAreas',
  postponedProjectAreas = 'postponedProjectAreas',
  replannedProjectAreas = 'replannedProjectAreas',
  presentRtuProjectAreas = 'presentRtuProjectAreas',
  presentProjectPins = 'presentProjectPins',
  presentRtuProjectPins = 'presentRtuProjectPins',
  projectCreation = 'projectCreation',
  roadNetworkArterial = 'roadNetworkArterial',
  roadSectionsSelection = 'roadSectionsSelection',
  selectionRadius = 'selectionRadius',
  streetTree = 'streetTree',
  track = 'track',
  trafficLight = 'trafficLight',
  tools = 'tools',
  unifiedSections = 'unifiedSection',
  circleComparison = 'circleComparison'
}